+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * EXOOutboundConnector
   * Fixed issue #821
+* O365OrgCustomizationSetting
+  * Fixes an issue where the resource was not being exported;
+* ODSettings
+  * Removed AD group guid dependency issue # 862
 * SPOTenantSettings
   * Parameter RequireAcceptingAccountMatchInvitedAccount
     is now deprecated (Issue #864)
@@ -12,8 +16,6 @@
   * Fixed issue # 866 changed domains to array
   * Fixed issue where trying to set anonymous link types if
     sharing not properly configured issue #865
-* ODSettings
-  * Removed AD group guid dependency issue # 862
 * MODULES
   * M365DSCUtil
     * Fixed an issue where function Test-M365DSCObjectHasProperty was missing
@@ -21,6 +23,9 @@
   * M365DSCReverse
     * Fixed an issue where passing in the file name and using the GUI for
       extraction did not store the file at the specified location (Issue #810)
+    * Fixed and issue where the -GenerateInfo parameter would always generate
+      a link to the same resource.
+    * Added current version module in the Export file.
 * DEPENDENCIES
   * AzureADPreview Updated to 2.0.2.119
   * DSCParser Updated to 1.3.0.0